+++ examples/simple_plan_with_report/variables.tf
@@ -1,4 +1,9 @@
 variable "env" {
-  type    = map(any)
-  default = {}
+  description = "Environment configuration map. Used to define environment-specific parameters like tags, resource names, and other settings"
+  type        = map(any)
+  default = {
+    Environment = "prod"
+    Owner       = "devops"
+    Terraform   = true
+  }
 }